# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: PhotoBomb API
  version: 1.0.0
  description: |
    Production-ready photo service API supporting uploads, metadata management, 
    search, albums, and sharing with presigned URLs and async processing.
  contact:
    name: PhotoBomb Team
    email: api@photobomb.example.com
servers:
  - url: https://api.photobomb.example.com/api/v1
    description: Production
  - url: https://api-staging.photobomb.example.com/api/v1
    description: Staging
  - url: http://localhost:8000/api/v1
    description: Local development
security:
  - bearerAuth: []
...
paths:
  /auth/register:
    post:
      tags: [auth]
      summary: Register new user
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password, full_name]
              properties:
                email:
                  type: string
                  format: email
                  example: user@example.com
                password:
                  type: string
                  format: password
                  minLength: 12
                  example: "MySecureP@ssw0rd123"
                full_name:
                  type: string
                  example: "Jane Doe"
      responses:
        '201':
          description: User registered successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          description: Email already registered
  /auth/login:
    post:
      tags: [auth]
      summary: Login with email/password
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
                  format: password
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          description: Invalid credentials
  /auth/refresh:
    post:
      tags: [auth]
      summary: Refresh access token
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refresh_token]
              properties:
                refresh_token:
                  type: string
                  example: "eyJhbGciOiJIUzI1NiIsInR5cCI6..."
      responses:
        '200':
          description: New access token
          content:
            application/json:
              schema:
                type: object
                properties:
                  access_token:
                    type: string
                  expires_in:
                    type: integer
                    example: 3600
        '401':
          description: Invalid or expired refresh token
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    internalApiKey:
      type: apiKey
      in: header
      name: X-Internal-API-Key
  schemas:
    AuthResponse:
      type: object
      properties:
        access_token:
          type: string
          example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
        refresh_token:
          type: string
          example: "eyJhbGciOiJIUzI1NiIsInR5cCI6..."
        expires_in:
          type: integer
          example: 3600
          description: Access token TTL in seconds
        user:
          $ref: '#/components/schemas/User'
    User:
      type: object
      properties:
        user_id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        full_name:
          type: string
        created_at:
          type: string
          format: date-time
        face_recognition_enabled:
          type: boolean
          description: User has opted in to face recognition
...
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
              details:
                type: array
                items:
                  type: object
                  properties:
                    field:
                      type: string
                    message:
                      type: string